update submodule and ignore file

+++ src/main.cpp
@@ -75,7 +75,7 @@ void computeParameterization() {
   }
 
   b(0) = bnd(0);
-  b(1) = bnd(round(bnd.size() / 2));
+  b(1) = bnd(static_cast<int>(round(bnd.size() / 2)));
   MatrixXd bc(2, 2);
   bc << 0, 0, 1, 0;
